requirements.txt
# Flask Framework
Flask

# Database
Flask-SQLAlchemy
Flask-Migrate

# Forms
Flask-WTF

# Template Rendering
Jinja2

# Optional for Bootstrap Integration
bootstrap-flask

# For running locally (Optional)
python-dotenv

# HTMX - Not a Python library, add via CDN in templates:
# <script src="https://unpkg.com/htmx.org"></script>
